standalone operation, without cljs build

+++ scripts/softlink_cljs.sh
@@ -6,13 +6,17 @@ if [ "$1" = "" ] ; then
 fi
 
 PUBLIC_DIR="$(dirname $0)/../public"
-# echo PUBLIC_DIR=${PUBLIC_DIR}
-# mkdir -p $PUBLIC_DIR
+
+# make the public dir if it doesn't already exist
+test ! -r ${PUBLIC_DIR} && mkdir -p ${PUBLIC_DIR}
+
 cd $PUBLIC_DIR
 echo PUBLIC_DIR=`pwd`
 
-for F in ../cljs-out/${1}/* ; do
-  echo "Linking: " $(basename ${F}) ;
-  rm -rf $(basename ${F}) ;
-  ln -s ${F} . ;
-done
+if [ -d ../cljs-out/${1} ] ; then
+  for F in ../cljs-out/${1}/* ; do
+    echo "Linking: " $(basename ${F}) ;
+    rm -rf $(basename ${F}) ;
+    ln -s ${F} . ;
+  done
+fi
